Create `SendBillingTransactionsService`
We move the work of sending billing transactions to the Charging Module (which involves iterating over the billing transactions, formatting and sending a request to the CM, and updating the transaction object to reflect this) into a new `SendBillingTransactionsService`
